Group support classes by type, move add-on classes and schemas into support, fix tests.
