+++ packages/add-ons/fractal-command-info/package.json
@@ -12,7 +12,7 @@
     "tildify": "^1.2.0"
   },
   "devDependencies": {
-    "@frctl/cli": "^2.0.0-beta.0",
+    "@frctl/support": "^2.0.0-beta.0",
     "@frctl/fractal": "^2.0.0-beta.0"
   }
 }
