+++ packages/internals/console/package.json
@@ -10,9 +10,10 @@
   "dependencies": {
     "@allmarkedup/climate": "^0.2.0",
     "clean-stack": "^1.3.0",
+    "common-tags": "^1.4.0",
+    "debug": "^3.0.0",
     "extract-stack": "^1.0.0",
     "indent-string": "^3.1.0",
-    "strip-indent": "^2.0.0",
-    "common-tags": "^1.4.0"
+    "strip-indent": "^2.0.0"
   }
 }
