+++ packages/internals/renderer/index.js
@@ -1,11 +1,7 @@
 const Renderer = require('./src/renderer');
-const adapterSchema = require('./src/adapter.schema');
 
 module.exports = function (fractal, opts = {}) {
   return new Renderer(fractal, opts);
 };
 
 module.exports.Renderer = Renderer;
-module.exports.schema = {
-  adapter: adapterSchema
-};
