+++ packages/internals/renderer/index.test.js
@@ -1,7 +1,6 @@
 const fractal = require('@frctl/fractal');
 const {expect} = require('../../../test/helpers');
 const Renderer = require('./src/renderer');
-const adapterSchema = require('./src/adapter.schema');
 const factory = require('.');
 
 describe('Renderer', () => {
@@ -12,10 +11,6 @@ describe('Renderer', () => {
 
     it('provides a reference to the Renderer class', () => {
       expect(factory.Renderer).to.equal(Renderer);
-    });
-
-    it('provides access to the adapter schema', () => {
-      expect(factory.schema.adapter).to.equal(adapterSchema);
     });
   });
 });
