+++ packages/internals/support/index.js
@@ -1,10 +1,12 @@
-module.exports.Collection = require('./src/collection');
-module.exports.ComponentCollection = require('./src/component-collection');
-module.exports.Entity = require('./src/entity');
-module.exports.File = require('./src/file');
-module.exports.FileCollection = require('./src/file-collection');
-module.exports.Component = require('./src/component');
-module.exports.Variant = require('./src/variant');
+module.exports.Collection = require('./src/collections/collection');
+module.exports.FileCollection = require('./src/collections/file-collection');
+module.exports.ComponentCollection = require('./src/collections/component-collection');
+module.exports.Entity = require('./src/entities/entity');
+module.exports.File = require('./src/entities/file');
+module.exports.Component = require('./src/entities/component');
+module.exports.Variant = require('./src/entities/variant');
+module.exports.Command = require('./src/add-ons/command');
+module.exports.Adapter = require('./src/add-ons/adapter');
 module.exports.Emitter = require('./src/emitter');
 module.exports.EmittingPromise = require('./src/emitting-promise');
 module.exports.Validator = require('./src/validator');
