+++ packages/internals/support/package.json
@@ -21,6 +21,7 @@
     "lodash": "^4.17.4",
     "multimatch": "^2.1.0",
     "slash": "^1.0.0",
+    "debug": "^3.0.0",
     "vinyl": "^2.1.0"
   }
 }
