+++ packages/internals/support/src/add-ons/adapter.schema.js
