+++ packages/internals/support/src/add-ons/command.schema.js
