+++ packages/internals/support/src/collections/collection.js
