+++ packages/internals/support/src/entities/entity.js
