+++ packages/internals/support/src/entities/file.js
