+++ packages/internals/support/src/entities/variant.js
