--- packages/toolkit/cli/index.js
@@ -1,3 +1,0 @@
-module.exports.schema = {
-  command: require('./src/command.schema')
-};
